XD-3412: More Refactoring
- Use common property names for the FTP and SFTP source
- Add regex option to FTP Source
- Reenable FTP integration test
