+++ ftp-source/.gitignore
@@ -1,0 +1,1 @@
+/.apt_generated/
